+++ src/main/java/de/samply/app/PatientTestDataGeneratorController.java
@@ -10,6 +10,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import utils.ProjectVersion;
 
@@ -23,12 +25,15 @@ public class PatientTestDataGeneratorController {
     private final String projectVersion = ProjectVersion.getProjectVersion();
     private final TestDataGenerator testDataGenerator;
     private final String testDataFilename;
+    private final int defaultNumberOfPatients;
 
     public PatientTestDataGeneratorController(
             TestDataGenerator testDataGenerator,
+            @Value(PatientTestDataGeneratorConst.DEFAULT_NUMBER_OF_PATIENTS_SV) int defaultNumberOfPatients,
             @Value(PatientTestDataGeneratorConst.TEST_DATA_FILENAME_SV) String testDataFilename) {
         this.testDataGenerator = testDataGenerator;
         this.testDataFilename = testDataFilename;
+        this.defaultNumberOfPatients = defaultNumberOfPatients;
     }
 
     @GetMapping(value = PatientTestDataGeneratorConst.INFO)
@@ -36,9 +41,12 @@ public class PatientTestDataGeneratorController {
         return new ResponseEntity<>(projectVersion, HttpStatus.OK);
     }
 
-    @GetMapping(value = PatientTestDataGeneratorConst.GENERATE)
-    public ResponseEntity<Resource> generate() throws IOException, TestDataGeneratorException {
-        return downloadDocument(testDataGenerator.generate());
+    @GetMapping(value = PatientTestDataGeneratorConst.GENERATE + PatientTestDataGeneratorConst.NUMBER_OF_PATIENTS)
+    public ResponseEntity<Resource> generate(@PathVariable (required = false) Integer numberOfPatients) throws IOException, TestDataGeneratorException {
+        if (numberOfPatients == null){
+            numberOfPatients = defaultNumberOfPatients;
+        }
+        return downloadDocument(testDataGenerator.generate(numberOfPatients));
     }
 
     private ResponseEntity<Resource> downloadDocument(Path filePath) throws IOException {